pyproject.toml
[project]
name = "datagouv_client"
version = "0.1.4.dev"
description = "Wrapper for the data.gouv.fr API"
authors = [{ name = "Etalab", email = "opendatateam@data.gouv.fr" }]
dependencies = [
    "httpx>=0.28.1,<1",
    "tenacity>=9.0.0,<10",
]
requires-python = ">=3.10,<3.14"
license = "MIT"
readme = "README.md"
keywords = ["api", "wrapper", "datagouv"]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1,<1",
    "pytest-httpx>=0.35.0,<1",
    "bumpx>=0.3.10",
    "ruff>=0.11.2",
]

[project.urls]
Source = "https://github.com/datagouv/datagouv_client"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
csv_detective = ["CHANGELOG.md", "LICENSE", "README.md"]

[tool.ruff]
lint = { extend-select = ["I"] } # ["I"] is to also sort imports with an isort rule
line-length = 100
